Update the spec of the builtins regarding the slet/sletu instructions.
In order to comply with the naming of other flavours of RISC-V
'set on less then' instructions we renamed the cv.slet/cv.sletu
instructions as cv.sle/cv.sleu in the LLVM support.
Aliases were added to provide backwords compatibility with
instructions cv.slet/cv.sletu.
We need for consistency to update the naming of the builtins
associated with such instructions.

Signed-off-by: Paolo Savini <paolo.savini@embecosm.com>